Update contact.html
yo yo i know there are some errors rn in the contact page, talked to formspree and gonna see if they can help fix it as im not sure what the error could be but i think it has to do with the formspree requiring an email but not requiring a name. im not sure why this is affecting the look but ill try to fix it.